Rename input file file
