+++ src/meshpy/four_c/beam_potential.py
@@ -30,7 +30,7 @@ interaction potentials."""
 
 from meshpy.core.boundary_condition import BoundaryCondition
 from meshpy.core.header_functions import get_yes_no
-from meshpy.four_c.inputfile import InputSection
+from meshpy.four_c.input_file import InputSection
 
 
 class BeamPotential:
