+++ tests/create_cubit_input.py
@@ -34,7 +34,7 @@ from cubitpy import CubitPy, cupy
 from cubitpy.mesh_creation_functions import create_brick, extrude_mesh_normal_to_surface
 
 from meshpy.core.conf import mpy
-from meshpy.four_c.inputfile import InputFile
+from meshpy.four_c.input_file import InputFile
 
 
 def create_tube_cubit_mesh(r, h, n_circumference, n_height):
