+++ tests/test_cosserat_curve.py
@@ -41,7 +41,7 @@ from meshpy.cosserat_curve.warping_along_cosserat_curve import (
     get_mesh_transformation,
     warp_mesh_along_curve,
 )
-from meshpy.four_c.inputfile import InputFile
+from meshpy.four_c.input_file import InputFile
 from meshpy.mesh_creation_functions.beam_basic_geometry import create_beam_mesh_helix
