+++ tests/test_mesh_creation_functions.py
@@ -42,7 +42,7 @@ from meshpy.core.material import MaterialEulerBernoulli, MaterialReissner
 from meshpy.core.mesh import Mesh
 from meshpy.core.node import NodeCosserat
 from meshpy.core.rotation import Rotation
-from meshpy.four_c.inputfile import InputFile
+from meshpy.four_c.input_file import InputFile
 from meshpy.mesh_creation_functions.beam_basic_geometry import (
     create_beam_mesh_arc_at_node,
     create_beam_mesh_arc_segment_2d,
