+++ tests/test_performance.py
@@ -38,7 +38,7 @@ from meshpy.core.element_beam import Beam3rHerm2Line3
 from meshpy.core.material import MaterialReissner
 from meshpy.core.mesh import Mesh
 from meshpy.core.rotation import Rotation
-from meshpy.four_c.inputfile import InputFile
+from meshpy.four_c.input_file import InputFile
 from meshpy.geometric_search.find_close_points import (
     FindClosePointAlgorithm,
     find_close_points,
